+++ stackview/_context.py
